arm64: dts: ti: k3-am654-base-board: Reserve memory for IPC between R5F cores
commit f82c5e0a8bc1311aee140bfed0888fc9a99afde0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a reserved memory node to reserve a portion of the DDR memory to be
used for performing inter-processor communication between all the MCU R5F
remote processors running RTOS on all the TI AM654 boards. This memory
shall be exercised only if the MCU R5FSS cluster is configured for Split
mode.  A single 1 MB of memory at 0xa2000000 is reserved for this purpose,
and this accounts for all the vrings and vring buffers between pair of
these R5F remote processors.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Lokesh Vutla <lokeshvutla@ti.com>
Link: https://lore.kernel.org/r/20201029033802.15366-5-s-anna@ti.com
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>