soc: ti: pruss: Refactor the CFG sub-module init
commit 8ac8812c87678bdb5fb68da85ec88fcc8fef7265 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

[ Upstream commit ea797f699440d85489dc839b28845434e350c917 ]

The CFG sub-module is not present on some earlier SoCs like the
DA850/OMAPL-138 in the TI Davinci family. Refactor out the CFG
sub-module parse and initialization logic into a separate function
to make it easier to add logic for the PRUSS IP on the above legacy
SoC families.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@oracle.com>
[s-anna@ti.com: cherry-pick commit 'ea797f699440' from v5.12]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>