PCI: endpoint: Add pci_epc_ops to map MSI IRQ
commit 5143deb339c68e34a4e50d43b3d9259d8b84ef78 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 87d5972e476f6c4e98a0abce713c54c6f40661b0 upstream.

Add pci_epc_ops to map physical address to MSI address and return MSI data.
The physical address is an address in the outbound region. This is required
to implement doorbell functionality of NTB (non-transparent bridge) wherein
EPC on either side of the interface (primary and secondary) can directly
write to the physical address (in outbound region) of the other interface
to ring doorbell using MSI.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Link: https://lore.kernel.org/r/20210201195809.7342-9-kishon@ti.com
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>