phy: core: Reword the comment specifying the units of max_link_rate to be Mbps
commit 2f10aee431696321f3ab02f3611a45c01829d615 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In some subsystems (eg. CAN, SPI), the max link rate supported can be less
than 1 Mbps and if the unit for max_link_rate is Mbps then it can't be
used. Therefore, leave the decision of units to be used, to the producer
and consumer.

Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>